pyproject.toml
[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["nonebot_plugin_whateat_pic"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
name = "nonebot-plugin-whateat-pic"
version = "1.5.3"
description = "基于Nonebot2的今天吃什么（离线版）"
authors = [
    {name = "Cvandia", email = "106718176+Cvandia@users.noreply.github.com"},
    {name = "Cvandia"},
]
dependencies = [
    "nonebot2>=2.0.0",
    "httpx>=0.19.0",
    "pillow>=10.4.0",
    "nonebot-plugin-alconna<1.0.0,>=0.53.1",
    "rich<14.0.0,>=13.9.4",
    "nonebot-plugin-apscheduler>=0.5.0",
]
requires-python = "<4,>=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/Cvandia/nonebot-plugin-whateat-pic"
repository = "https://github.com/Cvandia/nonebot-plugin-whateat-pic"

[dependency-groups]
dev = ["black<23.0.0,>=22.1.0", "ruff>=0.8.1", "pre-commit>=4.0.1"]
